refactor(ingest/snowflake): move oauth config into snowflake dir
