+++ metadata-ingestion/src/datahub/ingestion/source/snowflake/oauth_config.py
